[PIPE-10] File cleanups upon refactoring
