+++ .env.template
@@ -14,6 +14,7 @@
 # ==== [App] ====
 # MATVIEW_SQL_DIR has to be inside of the project (check the docker-compose file)
 MATVIEW_SQL_DIR=matview_sql
+PROJECT_LOG_DIR=./usaspending_api/logs
 
 # ==== [Django] ====
 # Use env var to toggle Django's DEBUG mode for local docker envs
