+++ requirements/requirements-app.txt
@@ -35,7 +35,6 @@ psycopg2-binary==2.8.*
 py-gfm==0.1.4
 pydantic[dotenv]==1.9.0
 pyspark==3.1.2
-python-dotenv==0.20.*
 python-json-logger==0.1.9
 requests==2.25.*
 retrying==1.3.3
