+++ usaspending_api/common/tests/test_etl.py
@@ -3,7 +3,7 @@ import pytest
 from collections import OrderedDict
 from pathlib import Path
 from psycopg2.sql import Identifier, Literal, SQL
-from usaspending_api.common.etl import (
+from usaspending_api.common.etl.postgres import (
     ETLDBLinkTable,
     ETLTable,
     ETLTemporaryTable,
