+++ usaspending_api/references/management/commands/load_city_county_state_code.py
@@ -6,7 +6,7 @@ from django.core.management.base import BaseCommand
 from django.db import connection, transaction
 from pathlib import Path
 from tempfile import TemporaryDirectory
-from usaspending_api.common.etl import ETLTable, ETLTemporaryTable
+from usaspending_api.common.etl.postgres import ETLTable, ETLTemporaryTable
 from usaspending_api.common.etl.postgres import operations
 from usaspending_api.common.helpers.sql_helpers import execute_dml_sql
 from usaspending_api.common.helpers.timing_helpers import ScriptTimer as Timer
